refactor: Split scale_up_service.py into modular strategy-based architecture
Breaks down 1,722-line god object into focused strategy modules:
- template_strategy.py (619 lines)
- scenario_strategy.py (350 lines)
- random_strategy.py (226 lines)
- common.py (321 lines)
- __init__.py (685 lines - orchestrator)

Total: 2,201 lines (modular overhead)

Each strategy is self-contained and independently testable.
Maintains 100% backward compatibility via re-exports.

Related: #531
